Replace PollEvent with rich events
Replace the GUI PollEvent method with the Event method that returns a
richer "event" object that can contain more information (E.g. modifiers
for a keypress) than just the two bits of information PollEvent
returned.
Define the events that we want. These are similar to, but distinct from,
the events that SDL returns: this is on purpose to stop SDL "leaking"
out of the GUI, so we translate from SDL event to internal event.
Modify the keyboard Scan() method to take a Keypress object instead of
just the scan code.