Clean up & DRY
Rename the test files so there's some namespacing.
Replace the literal Instruction{} definition in each test case with just
the opcode, and provide the Call() method to lookup & execute the
instruction from that.
Provide a testCases type with a Run method that will execute all of the
tests, instead of duplicating the loop into each Test*() method.