Completely refactor keyboard handling
Replace the overly complicated rune mapping with SDL->C= scancode, and
then have the keyboard convert *that* into a row/column. This means that
all of the normal keys work and things like shift are handled, but has
the current downside that any shifted keys on the input (E.g. !) can not
be typed, as that will be read as a shifted 1 which is a graphics
character on the PET keyboard.