Make shifted characters work
Work around absolute stupidity in SDL2 to extract the actual character
that was typed so that we can map characters on a modern keyboard like !
that would normally be shifted to keys on the PET keyboard where they
would not be, and where shifting them produces a *different* character.
This requires some SDL event shenanigans to get the text input event
which we need to get the actual keypress because SDL2 wont tell you in a
keydown even. Because reasons.
Once we have an actual character, we can map that rune to a PET scancode
and use that instead of the unshifted scancode.