DRY up bus access
Remove some duplicate implementations for reading & writing Words (as
two Bytes) by moving them up into the CPU bus as a set of functions that
can be embedded wherever they're needed. As a consequence the ReadWriter
interface is also moved from the emulator bus to the CPU, and the CPU
bus access functions are also re-organised slightly to fit the
composible interface model.